steppersync: Split steppersync code from stepcompress.c to new file
